I ran into an issue, added note about it in Readme
I expected tailwind-cascade to handle standard theme color names like 'primary', 'secondary', 'info' etc. But it doesn't. I was going to change the color definitions to include those but then I realized there are potential conflicts if people name their font sizes the same names (which you probably already thought of). But it is slightly unexpected behavior (as those color names are so common) so I figured there should be a note about it somewhere.